Moved status bar item creation to external file
